+++ lib/src/suite.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:tyto/src/models/case_result.dart';
 import 'package:tyto/src/ops_benchmark.dart';
 import 'package:tyto/src/report.dart';
+import 'package:tyto/src/utils.dart';
 
 /// A class that represents a suite of benchmarks.
 class Suite {
@@ -34,7 +35,9 @@ class Suite {
   int _currentWorstIndex = -1;
 
   /// Creates a new instance of [Suite].
-  Suite({this.measureDuration = const Duration(seconds: 1)});
+  Suite({
+      this.measureDuration = const Duration(seconds: 1),
+  });
 
   /// Adds a benchmark to the suite.
   void add(OpsBenchmarkBase benchmark) {
@@ -48,20 +51,31 @@ class Suite {
 
   /// Runs the benchmarks in the suite and generates the reports.
   Future<List<CaseResult>> run() async {
+    if(measureDuration.isNegative || measureDuration.inEffectiveSeconds == 0) {
+      throw ArgumentError('Measure duration must be greater than 0 seconds.');
+    }
     stdout.writeln('Running benchmarks...\n');
-    stdout.writeln('Each case will be run for $measureDuration seconds.\n');
+    stdout.writeln('Each case will be run for ${measureDuration.inEffectiveSeconds} seconds.\n');
     for (final benchmark in _benchmarks.indexed) {
       stdout.writeln('${benchmark.$2.name}:\n');
-      final score = await benchmark.$2.measure(measureDuration);
+      BenchmarkResult score;
+      try {
+        score = await benchmark.$2.measure(measureDuration);
+      } catch (e) {
+        score = BenchmarkResult.zero(benchmark.$2.name, benchmark.$2.group);
+        stdout.writeln(
+            '\tError: ${e.toString()}\n\tSkipping this benchmark.\n');
+        continue;
+      }
       _scores[benchmark.$1] = score;
       stdout.writeln(
-          '\t${score.avgScore} ops/sec ± ${score.stdDevPercentage.toStringAsFixed(2)}%\n');
+          '\t${score.avgScorePerSecond} ops/sec ± ${score.stdDevPercentage.toStringAsFixed(2)}%\n');
     }
     stdout.writeln('Finished ${_benchmarks.length} cases.');
     _currentBest = double.negativeInfinity;
     _currentWorst = double.infinity;
     for (final entry in _scores.entries) {
-      final score = entry.value.avgScore;
+      final score = entry.value.avgScorePerSecond;
 
       // Update the best and worst scores
       if (score > _currentBest) {
@@ -83,11 +97,12 @@ class Suite {
       final isBest = entry.key == _currentBestIndex;
       final isWorst = entry.key == _currentWorstIndex;
       final differenceFromBest =
-          ((_currentBest - score.avgScore) / _currentBest) * 100;
+          ((_currentBest - score.avgScorePerSecond) / _currentBest) * 100;
       results.add(CaseResult(
         group: benchmark.group,
         name: benchmark.name,
         avgScore: score.avgScore,
+        avgScorePerSecond: score.avgScorePerSecond,
         stdDevPercentage: score.stdDevPercentage,
         stdDev: score.stdDev,
         best: isBest,