+++ lib/src/utils.dart
@@ -1,0 +1,5 @@
+extension EffectiveInSeconds on Duration {
+
+  /// Converts the duration to seconds.
+  double get inEffectiveSeconds => inMicroseconds / Duration.microsecondsPerSecond;
+}