+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.1-dev.1
+
+- Remove the need of Isolates to run the onRun function in Benchmarks.
+- Add average score per second.
+- Add inEffectiveSeconds to prevent infinite values.
+
 ## 0.0.1-dev.0
 
 - Initial development version